imguiex cleanup
- Create Text, Inputs, Button partials
- Migrate ExcelCombos to ImGuiEx
- Get rid of the warning on imgui copy
- convert EColors to EzColors
