+++ ECommons/ImGuiMethods/ImGuiEx/Checkbox.cs
@@ -101,4 +101,22 @@ public static unsafe partial class ImGuiEx
         }
         return false;
     }
+
+    /// <summary>
+    /// Draws a checkbox that will be on the same line as previous if there is space, otherwise will move to the next line.
+    /// </summary>
+    /// <param name="label">Checkbox label</param>
+    /// <param name="v">Boolean to toggle</param>
+    /// <remarks><see cref="ImGui.SameLine()"/> does not need to be called just before using this.</remarks>
+    /// <returns></returns>
+    public static bool CheckboxWrapped(string label, ref bool v)
+    {
+        ImGui.SameLine();
+        var labelW = ImGui.CalcTextSize(label);
+        var finishPos = ImGui.GetCursorPosX() + labelW.X + ImGui.GetStyle().ItemSpacing.X + ImGui.GetStyle().ItemInnerSpacing.X + ImGui.GetStyle().FramePadding.Length() + ImGui.GetCursorStartPos().X;
+        if(finishPos >= ImGui.GetContentRegionMax().X)
+            ImGui.NewLine();
+
+        return ImGui.Checkbox(label, ref v);
+    }
 }
