+++ ECommons/ImGuiMethods/ImGuiEx/EzTable.cs
@@ -14,7 +14,7 @@ public static unsafe partial class ImGuiEx
     {
         if(!entries.Any())
         {
-            ImGuiEx.Text(EColor.RedBright, $"Table contains no elements!");
+            ImGuiEx.Text(EzColor.RedBright, $"Table contains no elements!");
             return;
         }
         var entriesArray = entries.ToArray();
